hp tune script fixing confidnet encoder  + inferen

+++ fd_shifts/exec.py
@@ -204,7 +204,7 @@ def test(cf):
         gpus=-1,
         logger=False,
         callbacks=[progress] + get_callbacks(cf),
-        precision=64,
+        precision=32,
         replace_sampler_ddp=False,
         # accelerator="ddp",
         accelerator=None,
